Windows port: ccl: narrowing casts and integer casts for MSVC
- all_broadcast, broadcast, all_gather (multicast/unicast), reduce_to_root:
  add static_cast<uint32_t> for DeviceAddr/size_t/int values placed into
  std::vector<uint32_t> compile-time/runtime-arg initializer lists, and for
  CoreCoord.x/.y fields (size_t) fed into those lists.
- common/uops/ccl_host_commands.cpp, common/uops/command_lowering.cpp,
  common/host/ccl_worker_builder.cpp, common/host/ccl_command_stream_builders.cpp,
  common/types/ccl_types_args_emitters.cpp:
  add explicit static_cast<uint32_t>/<uint8_t> in aggregate-initializer
  members that MSVC treats as narrowing (CclCommandAddrCircularBufferId,
  CclCommandAddrAbsoluteAddress, CclCommandCoreDescriptorTypeNocXY,
  CclCommandWaitValue, CclCommandAtomicInc, CclCommandInlineReadWrite,
  HostCclCommandNocTransferBurst, CclCommandTensor, Shape4D<uint32_t>);
  replace reinterpret_cast<uint32_t> on DeviceAddr with static_cast.
- all_to_all_combine, all_to_all_dispatch: cast size_t page sizes /
  mesh_view.num_rows()/num_cols() to uint32_t for CBDescriptor/
  CBFormatDescriptor fields and compile-time-arg vectors.

+++ ttnn/cpp/ttnn/operations/ccl/all_gather/device/all_gather_unicast_factory.cpp
@@ -570,8 +570,8 @@ AllGatherUnicastFactory::cached_program_t AllGatherUnicastFactory::create_at(
                     final_count,               // last-iteration slice length (even-ring split)
                     input_tile_id_start,       // local data: input page start
                     input_tile_id_end,         // local data: input page end
-                    barrier_sem.address(),     // barrier_sem L1 address
-                    data_valid_sem.address(),  // data_valid_sem L1 address
+                    static_cast<uint32_t>(barrier_sem.address()),     // barrier_sem L1 address
+                    static_cast<uint32_t>(data_valid_sem.address()),  // data_valid_sem L1 address
                 };
                 tt::tt_metal::SetRuntimeArgs(program, reader_kernel_id, {core}, reader_rt_args);
 
@@ -588,8 +588,8 @@ AllGatherUnicastFactory::cached_program_t AllGatherUnicastFactory::create_at(
                     final_start,                      // last-iteration slice start (even-ring split)
                     final_count,                      // last-iteration slice length (even-ring split)
                     do_local_write ? 1u : 0u,         // write local data into local output on iteration 0
-                    barrier_sem.address(),            // barrier_sem L1 address
-                    data_valid_sem.address(),         // data_valid_sem L1 address
+                    static_cast<uint32_t>(barrier_sem.address()),     // barrier_sem L1 address
+                    static_cast<uint32_t>(data_valid_sem.address()),  // data_valid_sem L1 address
                     (uint32_t)partner_core.x,         // barrier_sem target (neighbor partner core x)
                     (uint32_t)partner_core.y,         // barrier_sem target (neighbor partner core y)
                     (uint32_t)mirror_core.x,          // data_valid_sem target (neighbor mirror core x)
